Hack out extension updates
We can't replace the update check method as we could do before ES
modules, but we can re-define the updatesSupported property of the
extension manager on the extension manager directly, which then takes
precedent of the prototype definition.

I don't know how to undo this patch, but in case the extension is
disabled we can just call through to the prototype descriptor to get
the result of the original implementation.