Delete defined property in disable()
This removes the installed descriptor on the extension manager instance
and makes access to updatesEnabled call through to the original
prototype property again.

To simplify this and avoid that very concrete typescript types get
in our way of "delete" we abstract this into a PropertyManager class,
similar to the built-in InjectionManager which doesn't handle properties
unfortunately.